Add files via upload
The install.php script is a PHP script that can be used to install a MySQL database from a .sql file. It allows the user to enter the database hostname, username, password, and database name, as well as the path to the .sql file, and then automatically executes the queries in the .sql file to create the database tables and populate them with data.

The script includes a form for requesting the database information from the user and some basic CSS styles to make it mobile-friendly. It also includes error handling to display any errors that occur during the installation process.

The install.php script can be useful for automating the process of setting up a database for a web application or for quickly resetting the database to a known state during development or testing.